requirements.txt
pytest
pytest-xdist
playwright
pyyaml
requests